+++ octoprint_octorelay/migrations.py
@@ -10,8 +10,34 @@ def v0(settings, logger):
             after = { **before, "active": True }
             settings.set([index], after)
 
+def v1(settings, logger):
+    """Migration from v1 to v2"""
+    # Some settings were named in camelCase
+    replacements = {
+        "labelText": "label_text",
+        "cmdON": "cmd_on",
+        "cmdOFF": "cmd_off",
+        "autoONforPrint": "auto_on_before_print",
+        "autoOFFforPrint": "auto_off_after_print",
+        "autoOffDelay": "auto_off_delay",
+        "iconOn": "icon_on",
+        "iconOff": "icon_off",
+        "confirmOff": "confirm_off"
+    }
+    for index in ["r1", "r2", "r3", "r4", "r5", "r6", "r7", "r8"]:
+        before = settings.get([index])
+        after = {}
+        logger.debug(f"relay {index} stored settings: {before}")
+        for key, value in before.items():
+            if key in replacements:
+                after[replacements[key]] = value
+            else:
+                after[key] = value
+        logger.debug(f"replacing it with: {after}")
+        settings.set([index], after)
+
 # List of migration functions starting from v0->v1
-migrators = [ v0 ]
+migrators = [ v0, v1 ]
 
 def migrate(current: int, settings, logger):
     # Current version number corresponds to the list index to begin migrations from